Fix Anonka bot deployment issues
- Restore database connection retry logic (30 attempts with 2s delay) to handle MySQL initialization timing
- Fix main.py polling implementation with proper try/finally cleanup
- Update models.py to use BigInteger for all Telegram IDs to support large group chat IDs (e.g., -1003914225163)
- Update docker-compose.yml with MySQL native password authentication plugin for pymysql compatibility
- Fix MySQL port mapping from 3306 to 3307 to avoid conflicts with host MariaDB
- Add DB_PORT=3306 environment variable for bot container

These fixes resolve:
1. Database connection failures on startup
2. Out of range errors for negative group chat IDs
3. PyMySQL authentication errors with MySQL 8.0 caching_sha2_password
4. Bot polling not starting after connection

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Text, DateTime, JSON, ForeignKey
+from sqlalchemy import create_engine, Column, Integer, BigInteger, String, Text, DateTime, JSON, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
 from datetime import datetime
@@ -11,7 +11,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     username = Column(String(255))
     created_at = Column(DateTime, default=datetime.utcnow)
 
@@ -23,7 +23,7 @@ class Thread(Base):
 
     id = Column(Integer, primary_key=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    group_chat_id = Column(Integer, nullable=False)
+    group_chat_id = Column(BigInteger, nullable=False)
     topic_id = Column(Integer)
     created_at = Column(DateTime, default=datetime.utcnow)
 
@@ -36,7 +36,7 @@ class Message(Base):
 
     id = Column(Integer, primary_key=True)
     thread_id = Column(Integer, ForeignKey("threads.id"), nullable=False)
-    sender_telegram_id = Column(Integer, nullable=False)
+    sender_telegram_id = Column(BigInteger, nullable=False)
     sender_username = Column(String(255))
     message_text = Column(Text)
     attachments = Column(JSON)
@@ -51,7 +51,7 @@ class Response(Base):
 
     id = Column(Integer, primary_key=True)
     message_id = Column(Integer, ForeignKey("messages.id"), nullable=False)
-    responder_user_id = Column(Integer, nullable=False)
+    responder_user_id = Column(BigInteger, nullable=False)
     response_text = Column(Text)
     attachments = Column(JSON)
     created_at = Column(DateTime, default=datetime.utcnow)
